增加编辑状态
